Fixed project structure and removed unessary code
