+++ flight_passenger_map.py
@@ -1,0 +1,23 @@
+
+flight_mappings = {
+   "TS284": 169, "WG6146": 157, "TS890": 169, "F82100": 157, "WG378": 157,
+    "TS398": 169, "TS538": 169, "WG7188": 157, "CM484": 114, "WG4126": 157,
+    "AC944": 120, "AC1748": 144, "TS198": 169, "TS760": 169, "AC938": 262,
+    "AC954": 144, "EK244": 319, "AC962": 120, "AC948": 144, "AC1838": 144,
+    "WG5163": 157, "WG4134": 157, "WG7134": 157, "WG2189": 157, "AM681": 136,
+    "AC999": 144, "TS106": 169, "AC1004": 120, "TS938": 169, "TS498": 169,
+    "TS974": 169, "TS814": 169, "AC1884": 120, "LX87": 225, "AF345": 231,
+    "AC959": 120, "WG517": 157, "WG3106": 157, "AC844": 120, "AC874": 319,
+    "AC1792": 169, "OS74": 169, "AC995": 120, "AC832": 120, "TS196": 169,
+    "AC1275": 120, "AF347": 231, "KL672": 262, "AC866": 120, "TS110": 169,
+    "LH475": 261, "TS680": 169, "AC822": 120, "AC878": 120, "AC894": 144,
+    "AT209": 233, "BA94": 260, "QR764": 319, "AF4083": 123, "AC870": 319,
+    "S4328": 157, "AC876": 262, "TP254": 120, "AV201": 120, "RJ272": 233,
+    "WG604": 136, "WG525": 136, "WG2743": 136, "TS738": 169, "TS894": 169,
+    "TS356": 169, "AC922": 120, "WG434": 136, "AC1874": 120, "AC1450": 120,
+    "WG792": 136, "AC1325": 120, "AC920": 120, "WG537": 136, "WG428": 136,
+    "TS600": 169, "WG2789": 136, "AC1362": 120, "TS340": 169, "AF625": 231,
+    "AC5": 120, "AC1750": 120, "TS716": 169, "WG652": 136, "TS856": 169,
+    "AC781": 120, "TS602": 169, "AC892": 120, "TS252": 169, "AC50": 120,
+    "TK36": 260, "AC834": 120, "AC2076": 120, "AC96": 120, "AC2176": 120
+}
